Add files via upload
This is the ARP scanner lab solution that I came up with after some googling.
It may not appear to work as first. Through the googling, I found that people often ask for the interface to be specified and that it was defaulting to my VirtualBox NIC. Disabling the VirtualBox NIC solves this and will have the generated ARP table reflect the arp -a command. 
(You will need to do this if it appears as if no devices are on the generated arp table)